deep-research/requirements.txt
pydantic-ai
faiss-cpu
sentence-transformers 
python-dotenv 
numpy
python-dotenv